Updated Tor.framework dependencies: Tor 0.4.0.6 -> 0.4.2.7, OpenSSL 1.1.1d -> 1.1.1f, liblzma 5.2.4 -> 5.2.5

+++ Tor/TORLogging.m
@@ -156,7 +156,7 @@ static inline os_log_type_t TORLogTypeFromSeverity(int severity) {
     }
 }
 
-static void TORLogCallback(int severity, uint32_t domain, const char *msg) {
+static void TORLogCallback(int severity, uint64_t domain, const char *msg) {
     if (domain & LD_NOCB) {
         return;
     }
